# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Feedback service",
        "title": "Tourmate - Feedback Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/payments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a paginated list of payments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get all payments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Payment status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Payment method",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PaginationDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/callback-cancel/{id}": {
            "get": {
                "description": "Handles redirect or callback after canceled payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Callback after canceled payment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/callback-success/{id}": {
            "get": {
                "description": "Handles redirect or callback after successful payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Callback after successful payment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/customer/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of payments made by a specific customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get payments by user ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter property (e.g. date, price)",
                        "name": "filter_prop",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (ASC or DESC)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Payment status (e.g. pending, paid)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Payment Method (e.g. VTP, GHTK)",
                        "name": "method",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PaginationDataResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update payment details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Update a payment record",
                "parameters": [
                    {
                        "description": "UpdatePaymentRequest",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a single payment record by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get payment by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Payment"
                        }
                    },
                    "400": {
                        "description": "Invalid data. Please try again.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "401": {
                        "description": "You have no rights to access this action.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    },
                    "500": {
                        "description": "There is something wrong in the system during the process. Please try again later.",
                        "schema": {
                            "$ref": "#/definitions/response.MessageApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Payment": {
            "type": "object",
            "properties": {
                "created_date": {
                    "type": "string"
                },
                "customer_id": {
                    "type": "integer"
                },
                "invoice_id": {
                    "type": "integer"
                },
                "payment_id": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdatePaymentRequest": {
            "type": "object",
            "required": [
                "payment_id"
            ],
            "properties": {
                "method": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "response.MessageApiResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.PaginationDataResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "has_next": {
                    "type": "boolean"
                },
                "has_previous": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}